Clean error mess, add some copy
