+++ src/Platform/Windows/System/ErrorMessage.h
@@ -22,5 +22,7 @@
 
 namespace System {
 
+std::string lastErrorMessage();
+std::string errorMessage(int);
 
 }
