--- src/System/ErrorMessage.cpp
@@ -1,50 +1,0 @@
-// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
-// Copyright (c) 2016-2019, The Karbo developers
-//
-// This file is part of Karbo.
-//
-// Karbo is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Karbo is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with Karbo.  If not, see <http://www.gnu.org/licenses/>.
-
-#include "ErrorMessage.h"
-
-#ifndef NOMINMAX
-#define NOMINMAX
-#endif
-
-#include <cstddef>
-#include <windows.h>
-
-namespace System {
-
-std::string lastErrorMessage() {
-  return errorMessage(GetLastError());
-}
-
-std::string errorMessage(int error) {
-  struct Buffer {
-    ~Buffer() {
-      if (pointer != nullptr) {
-        LocalFree(pointer);
-      }
-    }
-
-    LPTSTR pointer = nullptr;
-  } buffer;
-
-  auto size = FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_ALLOCATE_BUFFER, nullptr, error,
-                            MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), reinterpret_cast<LPTSTR>(&buffer.pointer), 0, nullptr);
-  return "result=" + std::to_string(error) + ", " + std::string(buffer.pointer, size);
-}
-
-}
